requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
base58==2.1.1
bitarray==2.6.0
black==21.10b0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cytoolz==0.12.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.4.0
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
hexbytes==0.2.3
idna==3.3
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.10.0
lru-dict==1.1.8
Markdown==3.3.4
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
names==0.3.0
netaddr==0.8.0
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.3
parsimonious==0.8.1
pathspec==0.9.0
pkgutil_resolve_name==1.3.10
platformdirs==2.4.0
protobuf==3.20.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
rel==0.4.7
requests==2.26.0
requests-oauthlib==1.3.1
rlp==2.0.1
six==1.16.0
smmap==5.0.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
toolz==0.12.0
tqdm==4.64.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.10.0
web3==5.30.0
websocket-client==1.3.1
websockets==9.1
wget==3.2
yarl==1.8.1
zipp==3.8.1
